requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4
    # via virtualenv
cfgv==3.2.0
    # via pre-commit
click==7.1.2
    # via pip-tools
distlib==0.3.1
    # via virtualenv
filelock==3.0.12
    # via virtualenv
identify==1.5.12
    # via pre-commit
jinja2==2.11.3
    # via -r requirements.in
markupsafe==1.1.1
    # via
    #   -r requirements.in
    #   jinja2
mitogen==0.3.0rc1
    # via -r requirements.in
nodeenv==1.5.0
    # via pre-commit
pip-tools==5.5.0
    # via -r requirements.in
pre-commit==2.9.3
    # via -r requirements.in
pyyaml==5.3.1
    # via
    #   -r requirements.in
    #   pre-commit
six==1.15.0
    # via virtualenv
toml==0.10.2
    # via pre-commit
virtualenv==20.3.0
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# pip
